refactor(Ticket): Ticket.kt 패키지 변경
- 이전: entity
- 이후: model
